add configurable on-off for query statistics
