+++ hadoop/src/main/java/org/apache/carbondata/hadoop/CarbonInputFormat.java
@@ -48,9 +48,7 @@ import org.apache.carbondata.core.carbon.datastore.impl.btree.BlockBTreeLeafNode
 import org.apache.carbondata.core.carbon.metadata.schema.table.CarbonTable;
 import org.apache.carbondata.core.carbon.path.CarbonStorePath;
 import org.apache.carbondata.core.carbon.path.CarbonTablePath;
-import org.apache.carbondata.core.carbon.querystatistics.QueryStatistic;
-import org.apache.carbondata.core.carbon.querystatistics.QueryStatisticsConstants;
-import org.apache.carbondata.core.carbon.querystatistics.QueryStatisticsRecorder;
+import org.apache.carbondata.core.carbon.querystatistics.*;
 import org.apache.carbondata.core.constants.CarbonCommonConstants;
 import org.apache.carbondata.core.keygenerator.KeyGenException;
 import org.apache.carbondata.core.util.CarbonProperties;
@@ -461,8 +459,15 @@ public class CarbonInputFormat<T> extends FileInputFormat<Void, T> {
       FilterExpressionProcessor filterExpressionProcessor,
       AbsoluteTableIdentifier absoluteTableIdentifier, FilterResolverIntf resolver,
       String segmentId) throws IndexBuilderException, IOException {
-
-    QueryStatisticsRecorder recorder = new QueryStatisticsRecorder("");
+    String queryStatisticsRecorderInstanceType = CarbonProperties.getInstance()
+            .getProperty(CarbonCommonConstants.ENABLE_QUERY_STATISTICS,
+                    CarbonCommonConstants.ENABLE_QUERY_STATISTICS_DEFAULT);
+    QueryStatisticsRecorder recorder = null;
+    if (queryStatisticsRecorderInstanceType.equalsIgnoreCase("true")) {
+      recorder = new QueryStatisticsRecorderImpl("");
+    } else {
+      recorder = new QueryStatisticsRecorderDummy("");
+    }
     QueryStatistic statistic = new QueryStatistic();
     Map<String, AbstractIndex> segmentIndexMap =
         getSegmentAbstractIndexs(job, absoluteTableIdentifier, segmentId);
